Create README.md
This project is useful for tourists who are unfamiliar with the places where they The main purpose of “Travell Website” is instead of to provide the best traveling services to the customers and travel agents, we have expanded tours and travel administration strategy to provide an exploration platform where a tourist can find their trip spots according to their choices. This method further assists to promote reliable and fascinating tourism so that people can celebrate their vacations in their favourite places. 

Basically, the purpose of travel sites are to provide us a platform from where we can choose our travel destination along with an estimation of expensive Moreover, it guides us and helps us to do a complete verification and every other requirement with a safety security. Some points are given below :
To Create Web Based Application For our Organization. 
To Provide Search Facility For Customer. 
To Generate Different Types of Reports. 
To Provide the online Package Ticket Booking and online Payment 
Facility 
For Customer. 
To Provide package Details.